webapp/requirements.txt
Keras==2.2.4
Pillow==5.4.1
Django==2.1.5
tensorflow==1.12.0
requests
